REF: левая панель навигации описана теперь в корневой Acrivity
