+++ OTUSFilmCatalog/app/src/main/java/com/silicium/otusfilmcatalog/RootActivity.java
@@ -1,17 +1,25 @@
 package com.silicium.otusfilmcatalog;
 
+import android.content.DialogInterface;
 import android.os.Bundle;
+import android.view.MenuItem;
 import android.view.ViewGroup;
 import android.widget.FrameLayout;
 
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
+import androidx.appcompat.app.ActionBarDrawerToggle;
+import androidx.appcompat.app.AlertDialog;
 import androidx.appcompat.app.AppCompatActivity;
+import androidx.appcompat.widget.Toolbar;
 import androidx.constraintlayout.widget.ConstraintLayout;
+import androidx.core.view.GravityCompat;
+import androidx.drawerlayout.widget.DrawerLayout;
 import androidx.fragment.app.Fragment;
 import androidx.fragment.app.FragmentManager;
 import androidx.fragment.app.FragmentTransaction;
 
+import com.google.android.material.navigation.NavigationView;
 import com.silicium.otusfilmcatalog.logic.controller.MetricsStorage;
 import com.silicium.otusfilmcatalog.logic.model.FragmentWithCallback;
 import com.silicium.otusfilmcatalog.logic.model.IGotoFragmentCallback;
@@ -23,7 +31,7 @@ import com.silicium.otusfilmcatalog.ui.MainFragment;
 
 import static android.content.res.Configuration.ORIENTATION_PORTRAIT;
 
-public class RootActivity extends AppCompatActivity implements IGotoFragmentCallback, FragmentManager.OnBackStackChangedListener {
+public class RootActivity extends AppCompatActivity implements IGotoFragmentCallback, FragmentManager.OnBackStackChangedListener, NavigationView.OnNavigationItemSelectedListener  {
 
     @Override
     protected void onCreate(@Nullable Bundle savedInstanceState) {
@@ -38,6 +46,18 @@ public class RootActivity extends AppCompatActivity implements IGotoFragmentCall
 
         getSupportFragmentManager().addOnBackStackChangedListener(this);
         MetricsStorage.getMetricNotifier().notifyObserversDataChanged();
+
+        Toolbar toolbar = findViewById(R.id.toolbar);
+
+        DrawerLayout drawer = findViewById(R.id.drawer_layout);
+        ActionBarDrawerToggle toggle = new ActionBarDrawerToggle(
+                this, drawer, toolbar,
+                R.string.navigation_drawer_open, R.string.navigation_drawer_close);
+        drawer.addDrawerListener(toggle);
+        toggle.syncState();
+
+        NavigationView navigationView = findViewById(R.id.nav_view);
+        navigationView.setNavigationItemSelectedListener(this);
     }
 
     @Override
@@ -155,5 +175,37 @@ public class RootActivity extends AppCompatActivity implements IGotoFragmentCall
                 root_fragment.setLayoutParams(new ConstraintLayout.LayoutParams(ViewGroup.LayoutParams.MATCH_PARENT, ViewGroup.LayoutParams.MATCH_PARENT));
             }
         }
+    }
+
+    @Override
+    public boolean onNavigationItemSelected(@NonNull MenuItem item) {
+        int id = item.getItemId();
+
+        if (id == R.id.item_home)
+            gotoMainFragment();
+        else if (id == R.id.item_about)
+            gotoAboutFragment();
+        else if (id == R.id.item_exit) {
+            AlertDialog.Builder bld = new AlertDialog.Builder(this);
+            DialogInterface.OnClickListener exitDo =
+                    new DialogInterface.OnClickListener() {
+                        @Override
+                        public void onClick(DialogInterface dialog, int which) {
+                            System.exit(0);
+                        }
+                    };
+
+            bld.setMessage(R.string.exitDialogMessage);
+            bld.setTitle(R.string.exitDialogTitle);
+            bld.setNegativeButton(R.string.exitDialogNegativeAnswer, null);
+            bld.setPositiveButton(R.string.exitDialogPositiveAnswer, exitDo);
+            bld.setCancelable(false);
+            AlertDialog dialog = bld.create();
+            dialog.show();
+        }
+
+        DrawerLayout drawer = findViewById(R.id.drawer_layout);
+        drawer.closeDrawer(GravityCompat.START);
+        return true;
     }
 }
