+++ OTUSFilmCatalog/app/src/main/java/com/silicium/otusfilmcatalog/ui/AboutFragment.java
@@ -1,6 +1,5 @@
 package com.silicium.otusfilmcatalog.ui;
 
-import android.content.DialogInterface;
 import android.os.Bundle;
 import android.view.LayoutInflater;
 import android.view.MenuItem;
@@ -9,23 +8,16 @@ import android.view.ViewGroup;
 
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
-import androidx.appcompat.app.ActionBarDrawerToggle;
-import androidx.appcompat.app.AlertDialog;
-import androidx.appcompat.widget.Toolbar;
-import androidx.core.view.GravityCompat;
-import androidx.drawerlayout.widget.DrawerLayout;
 import androidx.fragment.app.Fragment;
 
 import com.google.android.material.bottomnavigation.BottomNavigationView;
-import com.google.android.material.navigation.NavigationView;
 import com.silicium.otusfilmcatalog.R;
 import com.silicium.otusfilmcatalog.logic.model.FragmentWithCallback;
 import com.silicium.otusfilmcatalog.ui.cuctomcomponents.UiComponents;
 
-public class AboutFragment extends FragmentWithCallback implements NavigationView.OnNavigationItemSelectedListener {
+public class AboutFragment extends FragmentWithCallback {
 
     public final static String FRAGMENT_TAG = AboutFragment.class.getSimpleName();
-    private View rootView;
 
     @Override
     @Nullable
@@ -39,8 +31,6 @@ public class AboutFragment extends FragmentWithCallback implements NavigationVie
     public void onViewCreated(@NonNull View view, @Nullable Bundle savedInstanceState) {
         super.onViewCreated(view, savedInstanceState);
 
-        rootView = view;
-
         final BottomNavigationView bnv = view.findViewById(R.id.bottom_about_navigation);
         bnv.setOnNavigationItemSelectedListener(new BottomNavigationView.OnNavigationItemSelectedListener() {
             @Override
@@ -52,48 +42,5 @@ public class AboutFragment extends FragmentWithCallback implements NavigationVie
 
         bnv.getMenu().getItem(0).setCheckable(false);
         bnv.getMenu().getItem(1).setCheckable(false);
-        bnv.getMenu().getItem(2).setCheckable(false);
-
-        Toolbar toolbar = view.findViewById(R.id.fragment_main_toolbar);
-
-        DrawerLayout drawer = view.findViewById(R.id.drawer_main_layout);
-        ActionBarDrawerToggle toggle = new ActionBarDrawerToggle(
-                getActivity(), drawer, toolbar,
-                R.string.navigation_drawer_open, R.string.navigation_drawer_close);
-        drawer.addDrawerListener(toggle);
-        toggle.syncState();
-
-        NavigationView navigationView = view.findViewById(R.id.nav_view);
-        navigationView.setNavigationItemSelectedListener(this);
-    }
-
-    @Override
-    public boolean onNavigationItemSelected(@NonNull MenuItem item) {
-        int id = item.getItemId();
-
-        if (id == R.id.item_home)
-            gotoFragmentCallback.gotoMainFragment();
-        else if (id == R.id.item_exit) {
-            AlertDialog.Builder bld = new AlertDialog.Builder(rootView.getContext());
-            DialogInterface.OnClickListener exitDo =
-                    new DialogInterface.OnClickListener() {
-                        @Override
-                        public void onClick(DialogInterface dialog, int which) {
-                            System.exit(0);
-                        }
-                    };
-
-            bld.setMessage(R.string.exitDialogMessage);
-            bld.setTitle(R.string.exitDialogTitle);
-            bld.setNegativeButton(R.string.exitDialogNegativeAnswer, null);
-            bld.setPositiveButton(R.string.exitDialogPositiveAnswer, exitDo);
-            bld.setCancelable(false);
-            AlertDialog dialog = bld.create();
-            dialog.show();
-        }
-
-        DrawerLayout drawer = rootView.findViewById(R.id.drawer_main_layout);
-        drawer.closeDrawer(GravityCompat.START);
-        return true;
-    }
+        bnv.getMenu().getItem(2).setCheckable(false);    }
 }
